requirements.txt
async-timeout==4.0.3
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
redis==5.0.1
Werkzeug==3.0.1
pyjwt==2.8.0
flask_jwt_extended==4.6.0